refactor: split admin_handler.go into sub-domain files
Final layer of the admin god-file split. Query-param helpers
(queryInt, queryPerPage, queryBool) and paginatedResponse stay in
core admin_handler.go so all sub-domain handler files share them
package-locally.

No behavior change.
