+++ backend/internal/handler/admin_handler.go
@@ -1,17 +1,12 @@
 package handler
 
 import (
-	"encoding/json"
-	"errors"
 	"log/slog"
 	"net/http"
 	"strconv"
-	"time"
 
-	"github.com/go-chi/chi/v5"
 	"github.com/go-playground/validator/v10"
 
-	"github.com/jaochai/pixlinks/backend/internal/middleware"
 	"github.com/jaochai/pixlinks/backend/internal/service"
 )
 
@@ -27,263 +22,6 @@ func NewAdminHandler(adminService *service.AdminService, logger *slog.Logger) *A
 		validate:     newValidator(),
 		logger:       logger,
 	}
-}
-
-func (h *AdminHandler) ListCustomers(w http.ResponseWriter, r *http.Request) {
-	search := r.URL.Query().Get("search")
-	plan := r.URL.Query().Get("plan")
-	status := r.URL.Query().Get("status")
-	page := queryInt(r, "page", 1)
-	perPage := queryPerPage(r)
-
-	customers, total, err := h.adminService.ListCustomers(r.Context(), search, plan, status, page, perPage)
-	if err != nil {
-		if errors.Is(err, service.ErrInvalidPlan) {
-			ErrorJSON(w, http.StatusBadRequest, "invalid plan filter")
-			return
-		}
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to list customers", err)
-		return
-	}
-
-	// Strip sensitive data — admin list should not expose API keys or password hashes
-	for _, c := range customers {
-		c.APIKey = ""
-		c.PasswordHash = ""
-	}
-
-	totalPages := total / perPage
-	if total%perPage != 0 {
-		totalPages++
-	}
-
-	JSON(w, http.StatusOK, PaginatedResponse{
-		Data:       customers,
-		Total:      total,
-		Page:       page,
-		PerPage:    perPage,
-		TotalPages: totalPages,
-	})
-}
-
-func (h *AdminHandler) GetCustomerDetail(w http.ResponseWriter, r *http.Request) {
-	customerID := chi.URLParam(r, "id")
-
-	detail, err := h.adminService.GetCustomerDetail(r.Context(), customerID)
-	if err != nil {
-		if errors.Is(err, service.ErrCustomerNotFound) {
-			ErrorJSON(w, http.StatusNotFound, "customer not found")
-			return
-		}
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to get customer detail", err)
-		return
-	}
-
-	// Strip sensitive data
-	detail.Customer.PasswordHash = ""
-
-	JSON(w, http.StatusOK, APIResponse{Data: detail})
-}
-
-func (h *AdminHandler) ChangePlan(w http.ResponseWriter, r *http.Request) {
-	adminID := middleware.GetCustomerID(r.Context())
-	customerID := chi.URLParam(r, "id")
-
-	var input struct {
-		Plan string `json:"plan" validate:"required"`
-	}
-	if err := json.NewDecoder(r.Body).Decode(&input); err != nil {
-		ErrorJSON(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-	if err := h.validate.Struct(input); err != nil {
-		ErrorJSON(w, http.StatusBadRequest, FormatValidationErrors(err))
-		return
-	}
-
-	if err := h.adminService.ChangePlan(r.Context(), adminID, customerID, input.Plan); err != nil {
-		switch {
-		case errors.Is(err, service.ErrCustomerNotFound):
-			ErrorJSON(w, http.StatusNotFound, "customer not found")
-		case errors.Is(err, service.ErrInvalidPlan):
-			ErrorJSON(w, http.StatusBadRequest, "invalid plan")
-		default:
-			ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to change plan", err)
-		}
-		return
-	}
-
-	JSON(w, http.StatusOK, APIResponse{Message: "plan updated"})
-}
-
-func (h *AdminHandler) SuspendCustomer(w http.ResponseWriter, r *http.Request) {
-	adminID := middleware.GetCustomerID(r.Context())
-	customerID := chi.URLParam(r, "id")
-
-	if err := h.adminService.SuspendCustomer(r.Context(), adminID, customerID); err != nil {
-		switch {
-		case errors.Is(err, service.ErrAdminSelfSuspend):
-			ErrorJSON(w, http.StatusBadRequest, "cannot suspend your own account")
-		case errors.Is(err, service.ErrCustomerNotFound):
-			ErrorJSON(w, http.StatusNotFound, "customer not found")
-		default:
-			ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to suspend customer", err)
-		}
-		return
-	}
-
-	JSON(w, http.StatusOK, APIResponse{Message: "customer suspended"})
-}
-
-func (h *AdminHandler) ActivateCustomer(w http.ResponseWriter, r *http.Request) {
-	adminID := middleware.GetCustomerID(r.Context())
-	customerID := chi.URLParam(r, "id")
-
-	if err := h.adminService.ActivateCustomer(r.Context(), adminID, customerID); err != nil {
-		if errors.Is(err, service.ErrCustomerNotFound) {
-			ErrorJSON(w, http.StatusNotFound, "customer not found")
-			return
-		}
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to activate customer", err)
-		return
-	}
-
-	JSON(w, http.StatusOK, APIResponse{Message: "customer activated"})
-}
-
-func (h *AdminHandler) GrantCredits(w http.ResponseWriter, r *http.Request) {
-	adminID := middleware.GetCustomerID(r.Context())
-	customerID := chi.URLParam(r, "id")
-
-	var input service.GrantCreditsInput
-	if err := json.NewDecoder(r.Body).Decode(&input); err != nil {
-		ErrorJSON(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-	if err := h.validate.Struct(input); err != nil {
-		ErrorJSON(w, http.StatusBadRequest, FormatValidationErrors(err))
-		return
-	}
-
-	grant, err := h.adminService.GrantCredits(r.Context(), adminID, customerID, input)
-	if err != nil {
-		if errors.Is(err, service.ErrCustomerNotFound) {
-			ErrorJSON(w, http.StatusNotFound, "customer not found")
-			return
-		}
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to grant credits", err)
-		return
-	}
-
-	JSON(w, http.StatusCreated, APIResponse{Data: grant})
-}
-
-func (h *AdminHandler) GetPlatformOverview(w http.ResponseWriter, r *http.Request) {
-	stats, err := h.adminService.GetPlatformOverview(r.Context())
-	if err != nil {
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to get platform overview", err)
-		return
-	}
-
-	JSON(w, http.StatusOK, APIResponse{Data: stats})
-}
-
-func (h *AdminHandler) GetRevenueChart(w http.ResponseWriter, r *http.Request) {
-	days := queryInt(r, "days", 30)
-
-	chart, err := h.adminService.GetRevenueChart(r.Context(), days)
-	if err != nil {
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to get revenue chart", err)
-		return
-	}
-
-	JSON(w, http.StatusOK, APIResponse{Data: chart})
-}
-
-func (h *AdminHandler) GetGrowthChart(w http.ResponseWriter, r *http.Request) {
-	days := queryInt(r, "days", 30)
-
-	chart, err := h.adminService.GetGrowthChart(r.Context(), days)
-	if err != nil {
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to get growth chart", err)
-		return
-	}
-
-	JSON(w, http.StatusOK, APIResponse{Data: chart})
-}
-
-func (h *AdminHandler) ListPurchases(w http.ResponseWriter, r *http.Request) {
-	status := r.URL.Query().Get("status")
-	page := queryInt(r, "page", 1)
-	perPage := queryPerPage(r)
-
-	purchases, total, err := h.adminService.ListAllPurchases(r.Context(), status, page, perPage)
-	if err != nil {
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to list purchases", err)
-		return
-	}
-
-	totalPages := total / perPage
-	if total%perPage != 0 {
-		totalPages++
-	}
-
-	JSON(w, http.StatusOK, PaginatedResponse{
-		Data:       purchases,
-		Total:      total,
-		Page:       page,
-		PerPage:    perPage,
-		TotalPages: totalPages,
-	})
-}
-
-func (h *AdminHandler) ListSubscriptions(w http.ResponseWriter, r *http.Request) {
-	status := r.URL.Query().Get("status")
-	page := queryInt(r, "page", 1)
-	perPage := queryPerPage(r)
-
-	subs, total, err := h.adminService.ListAllSubscriptions(r.Context(), status, page, perPage)
-	if err != nil {
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to list subscriptions", err)
-		return
-	}
-
-	totalPages := total / perPage
-	if total%perPage != 0 {
-		totalPages++
-	}
-
-	JSON(w, http.StatusOK, PaginatedResponse{
-		Data:       subs,
-		Total:      total,
-		Page:       page,
-		PerPage:    perPage,
-		TotalPages: totalPages,
-	})
-}
-
-func (h *AdminHandler) ListCreditGrants(w http.ResponseWriter, r *http.Request) {
-	page := queryInt(r, "page", 1)
-	perPage := queryPerPage(r)
-
-	grants, total, err := h.adminService.ListCreditGrants(r.Context(), page, perPage)
-	if err != nil {
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to list credit grants", err)
-		return
-	}
-
-	totalPages := total / perPage
-	if total%perPage != 0 {
-		totalPages++
-	}
-
-	JSON(w, http.StatusOK, PaginatedResponse{
-		Data:       grants,
-		Total:      total,
-		Page:       page,
-		PerPage:    perPage,
-		TotalPages: totalPages,
-	})
 }
 
 func queryInt(r *http.Request, key string, defaultVal int) int {
@@ -328,261 +66,4 @@ func (h *AdminHandler) paginatedResponse(w http.ResponseWriter, data interface{}
 		PerPage:    perPage,
 		TotalPages: totalPages,
 	})
-}
-
-// F1: Sale Pages
-
-func (h *AdminHandler) ListSalePages(w http.ResponseWriter, r *http.Request) {
-	search := r.URL.Query().Get("search")
-	customerID := r.URL.Query().Get("customer_id")
-	published := queryBool(r, "published")
-	page := queryInt(r, "page", 1)
-	perPage := queryPerPage(r)
-
-	pages, total, err := h.adminService.ListAllSalePages(r.Context(), search, customerID, published, page, perPage)
-	if err != nil {
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to list sale pages", err)
-		return
-	}
-
-	h.paginatedResponse(w, pages, total, page, perPage)
-}
-
-func (h *AdminHandler) GetSalePageDetail(w http.ResponseWriter, r *http.Request) {
-	id := chi.URLParam(r, "id")
-
-	detail, err := h.adminService.GetSalePageDetail(r.Context(), id)
-	if err != nil {
-		if errors.Is(err, service.ErrSalePageNotFound) {
-			ErrorJSON(w, http.StatusNotFound, "sale page not found")
-			return
-		}
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to get sale page detail", err)
-		return
-	}
-
-	JSON(w, http.StatusOK, APIResponse{Data: detail})
-}
-
-func (h *AdminHandler) DisableSalePage(w http.ResponseWriter, r *http.Request) {
-	adminID := middleware.GetCustomerID(r.Context())
-	id := chi.URLParam(r, "id")
-
-	if err := h.adminService.DisableSalePage(r.Context(), adminID, id); err != nil {
-		if errors.Is(err, service.ErrSalePageNotFound) {
-			ErrorJSON(w, http.StatusNotFound, "sale page not found")
-			return
-		}
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to disable sale page", err)
-		return
-	}
-
-	JSON(w, http.StatusOK, APIResponse{Message: "sale page disabled"})
-}
-
-func (h *AdminHandler) EnableSalePage(w http.ResponseWriter, r *http.Request) {
-	adminID := middleware.GetCustomerID(r.Context())
-	id := chi.URLParam(r, "id")
-
-	if err := h.adminService.EnableSalePage(r.Context(), adminID, id); err != nil {
-		if errors.Is(err, service.ErrSalePageNotFound) {
-			ErrorJSON(w, http.StatusNotFound, "sale page not found")
-			return
-		}
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to enable sale page", err)
-		return
-	}
-
-	JSON(w, http.StatusOK, APIResponse{Message: "sale page enabled"})
-}
-
-func (h *AdminHandler) DeleteSalePage(w http.ResponseWriter, r *http.Request) {
-	adminID := middleware.GetCustomerID(r.Context())
-	id := chi.URLParam(r, "id")
-
-	if err := h.adminService.DeleteSalePage(r.Context(), adminID, id); err != nil {
-		if errors.Is(err, service.ErrSalePageNotFound) {
-			ErrorJSON(w, http.StatusNotFound, "sale page not found")
-			return
-		}
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to delete sale page", err)
-		return
-	}
-
-	JSON(w, http.StatusOK, APIResponse{Message: "sale page deleted"})
-}
-
-// F2: Pixels
-
-func (h *AdminHandler) ListPixels(w http.ResponseWriter, r *http.Request) {
-	search := r.URL.Query().Get("search")
-	customerID := r.URL.Query().Get("customer_id")
-	active := queryBool(r, "active")
-	page := queryInt(r, "page", 1)
-	perPage := queryPerPage(r)
-
-	pixels, total, err := h.adminService.ListAllPixels(r.Context(), search, customerID, active, page, perPage)
-	if err != nil {
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to list pixels", err)
-		return
-	}
-
-	h.paginatedResponse(w, pixels, total, page, perPage)
-}
-
-func (h *AdminHandler) GetPixelDetail(w http.ResponseWriter, r *http.Request) {
-	id := chi.URLParam(r, "id")
-
-	detail, err := h.adminService.GetPixelDetail(r.Context(), id)
-	if err != nil {
-		if errors.Is(err, service.ErrPixelNotFound) {
-			ErrorJSON(w, http.StatusNotFound, "pixel not found")
-			return
-		}
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to get pixel detail", err)
-		return
-	}
-
-	JSON(w, http.StatusOK, APIResponse{Data: detail})
-}
-
-func (h *AdminHandler) DisablePixel(w http.ResponseWriter, r *http.Request) {
-	adminID := middleware.GetCustomerID(r.Context())
-	id := chi.URLParam(r, "id")
-
-	if err := h.adminService.DisablePixel(r.Context(), adminID, id); err != nil {
-		if errors.Is(err, service.ErrPixelNotFound) {
-			ErrorJSON(w, http.StatusNotFound, "pixel not found")
-			return
-		}
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to disable pixel", err)
-		return
-	}
-
-	JSON(w, http.StatusOK, APIResponse{Message: "pixel disabled"})
-}
-
-func (h *AdminHandler) EnablePixel(w http.ResponseWriter, r *http.Request) {
-	adminID := middleware.GetCustomerID(r.Context())
-	id := chi.URLParam(r, "id")
-
-	if err := h.adminService.EnablePixel(r.Context(), adminID, id); err != nil {
-		if errors.Is(err, service.ErrPixelNotFound) {
-			ErrorJSON(w, http.StatusNotFound, "pixel not found")
-			return
-		}
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to enable pixel", err)
-		return
-	}
-
-	JSON(w, http.StatusOK, APIResponse{Message: "pixel enabled"})
-}
-
-// F3: Replays
-
-func (h *AdminHandler) ListReplays(w http.ResponseWriter, r *http.Request) {
-	status := r.URL.Query().Get("status")
-	customerID := r.URL.Query().Get("customer_id")
-	page := queryInt(r, "page", 1)
-	perPage := queryPerPage(r)
-
-	sessions, total, err := h.adminService.ListAllReplaySessions(r.Context(), status, customerID, page, perPage)
-	if err != nil {
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to list replays", err)
-		return
-	}
-
-	h.paginatedResponse(w, sessions, total, page, perPage)
-}
-
-func (h *AdminHandler) GetReplayDetail(w http.ResponseWriter, r *http.Request) {
-	id := chi.URLParam(r, "id")
-
-	detail, err := h.adminService.GetReplayDetail(r.Context(), id)
-	if err != nil {
-		if errors.Is(err, service.ErrReplayNotFound) {
-			ErrorJSON(w, http.StatusNotFound, "replay session not found")
-			return
-		}
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to get replay detail", err)
-		return
-	}
-
-	JSON(w, http.StatusOK, APIResponse{Data: detail})
-}
-
-func (h *AdminHandler) CancelReplay(w http.ResponseWriter, r *http.Request) {
-	adminID := middleware.GetCustomerID(r.Context())
-	id := chi.URLParam(r, "id")
-
-	if err := h.adminService.CancelReplay(r.Context(), adminID, id); err != nil {
-		if errors.Is(err, service.ErrReplayNotFound) {
-			ErrorJSON(w, http.StatusNotFound, "replay session not found")
-			return
-		}
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to cancel replay", err)
-		return
-	}
-
-	JSON(w, http.StatusOK, APIResponse{Message: "replay cancelled"})
-}
-
-// F4: Events
-
-func (h *AdminHandler) ListEvents(w http.ResponseWriter, r *http.Request) {
-	customerID := r.URL.Query().Get("customer_id")
-	pixelID := r.URL.Query().Get("pixel_id")
-	eventName := r.URL.Query().Get("event_name")
-	page := queryInt(r, "page", 1)
-	perPage := queryPerPage(r)
-
-	events, total, err := h.adminService.ListAllEvents(r.Context(), customerID, pixelID, eventName, page, perPage)
-	if err != nil {
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to list events", err)
-		return
-	}
-
-	h.paginatedResponse(w, events, total, page, perPage)
-}
-
-func (h *AdminHandler) GetEventStats(w http.ResponseWriter, r *http.Request) {
-	hours := queryInt(r, "hours", 24)
-
-	stats, err := h.adminService.GetEventStats(r.Context(), hours)
-	if err != nil {
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to get event stats", err)
-		return
-	}
-
-	JSON(w, http.StatusOK, APIResponse{Data: stats})
-}
-
-// F5: Audit Log
-
-func (h *AdminHandler) ListAuditLog(w http.ResponseWriter, r *http.Request) {
-	adminID := r.URL.Query().Get("admin_id")
-	action := r.URL.Query().Get("action")
-	targetCustomerID := r.URL.Query().Get("target_customer_id")
-	page := queryInt(r, "page", 1)
-	perPage := queryPerPage(r)
-
-	var from, to *time.Time
-	if v := r.URL.Query().Get("from"); v != "" {
-		if t, err := time.Parse(time.RFC3339, v); err == nil {
-			from = &t
-		}
-	}
-	if v := r.URL.Query().Get("to"); v != "" {
-		if t, err := time.Parse(time.RFC3339, v); err == nil {
-			to = &t
-		}
-	}
-
-	entries, total, err := h.adminService.ListAuditLogs(r.Context(), adminID, action, targetCustomerID, from, to, page, perPage)
-	if err != nil {
-		ErrorJSONWithLog(w, r, h.logger, http.StatusInternalServerError, "failed to list audit log", err)
-		return
-	}
-
-	h.paginatedResponse(w, entries, total, page, perPage)
 }
